+++ test/dummy/app/controllers/application_controller.rb
@@ -1,4 +1,6 @@
 class ApplicationController < ActionController::Base
+  include RecordingStudio::UsesDefaultLayout
+
   # Only allow modern browsers supporting webp images, web push, badges, import maps, CSS nesting, and CSS :has.
   allow_browser versions: :modern
 
@@ -13,7 +15,7 @@ class ApplicationController < ActionController::Base
   private
 
   def application_layout
-    devise_controller? ? "application" : "flat_pack_sidebar"
+    devise_controller? ? "application" : "recording_studio/default_layout"
   end
 
   def set_current_actor