+++ test/dummy/app/models/workspace.rb
@@ -1,3 +1,4 @@
 class Workspace < ApplicationRecord
-	recording_studio_recordable label: "Workspace", plural_label: "Workspaces", root: true
+  recording_studio_recordable label: "Workspace", plural_label: "Workspaces", root: true
+  RecordingStudio.enable_capability(:accessible, on: self) if defined?(RecordingStudioAccessible)
 end