+++ test/dummy/test/test_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+ENV["RAILS_ENV"] ||= "test"
+
+require_relative "../config/environment"
+require "rails/test_help"
+require "devise/test/integration_helpers"